FEAT: Add CRUD for cards and move models to folder

+++ infrastructure/api/finance_control/views/user_view.py
@@ -2,7 +2,8 @@ from rest_framework import status
 from rest_framework.views import APIView
 from rest_framework.response import Response
 
-from finance_control.repositories.repository_user import UserRepository
+from infrastructure.api.finance_control.repositories.users_repository import UserRepository
+from infrastructure.api.finance_control.serializers.users_serializer import *
 
 from dataclasses import asdict
 
@@ -27,27 +28,30 @@ class UsersView(APIView):
     - last_name: Sobrenome do usuário (opcional)
     """
     def post(self, request):
-        username = request.data.get('username')
-        password = request.data.get('password')
-        email = request.data.get('email')
-        first_name = request.data.get('first_name')
-        last_name = request.data.get('last_name')
-
+        serializer = UserCreateSerializer(data=request.data)
+        if not serializer.is_valid():
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        
         repository = UserRepository()
-        response = repository.create_user(username, password, email, first_name, last_name)
+        response = repository.create_user(serializer.validated_data)
         return Response(asdict(response), status=status.HTTP_201_CREATED)
     
     def put(self, request, pk):
-        username = request.data.get('username')
-        email = request.data.get('email')
-        first_name = request.data.get('first_name')
-        last_name = request.data.get('last_name')
-    
+        if not pk:
+            return Response({'error': 'ID do usuário não informado'}, status=status.HTTP_400_BAD_REQUEST)
+        
+        serializer = UserUpdateSerializer(data=request.data)
+        if not serializer.is_valid():
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        
         repository = UserRepository()
-        response = repository.update_user(pk, username, email, first_name, last_name)
+        response = repository.update_user(pk, serializer.validated_data)
         return Response(asdict(response), status=status.HTTP_200_OK)
 
     def delete(self, request, pk):
+        if not pk:
+            return Response({'error': 'ID do usuário não informado'}, status=status.HTTP_400_BAD_REQUEST)
+        
         repository = UserRepository()
         repository.delete_user(pk)
         return Response(status=status.HTTP_204_NO_CONTENT)
